requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.6.1
supabase==2.3.0
python-dotenv==1.0.1
apify-client==1.6.1
